% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\docType{data}
\name{cate_path}
\alias{cate_path}
\title{Category hierarchy informarion of CompEngine database.}
\format{An object of class \code{list} of length 187.}
\source{
\url{https://www.comp-engine.org/#!browse}
}
\usage{
cate_path
}
\description{
A list containing the category hierarchy chain of CompEngine data.
Each list contains subcategories of the first category listed (the name of the list).
}
\keyword{datasets}
