% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category_scraping.R
\name{category_scraping}
\alias{category_scraping}
\title{Download category information}
\usage{
category_scraping()
}
\description{
Download category information
}
\examples{
cate_path <- category_scraping()
}
\author{
Yangzhuoran Yang
}
