% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\docType{data}
\name{cets}
\alias{cets}
\alias{cets1}
\alias{cets2}
\alias{cets3}
\alias{cets4}
\alias{cets5}
\alias{cets6}
\alias{cets7}
\alias{cets8}
\alias{cets9}
\alias{cets10}
\alias{cets11}
\alias{cets12}
\alias{cets13}
\alias{cets14}
\alias{cets15}
\alias{cets16}
\alias{cets17}
\alias{cets18}
\alias{cets19}
\alias{cets20}
\alias{cets21}
\alias{cets22}
\alias{cets23}
\alias{cets24}
\alias{cets25}
\title{Time series data from CompEngine database.}
\format{An object of class \code{list} of length 1000 except for the last set.
The length of the last set depends on the current size of CompEngine database.}
\source{
\url{https://www.comp-engine.org/#!browse}
}
\usage{
cets1

cets2

cets3

cets4

cets5

cets6

cets7

cets8

cets9

cets10

cets11

cets12

cets13

cets14

cets15

cets16

cets17

cets18

cets19

cets20

cets21

cets22

cets23

cets24

cets25
}
\description{
Lists containing time series of type "numeric". Attributes attached. See also \code{?meta}.
To extract specific type of time series, see \code{?get_cets}.
}
\keyword{datasets}
