% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cets.R
\docType{data}
\name{cets10}
\alias{cets10}
\title{Data from CompEngine}
\format{An object of class \code{list} of length 1000.}
\usage{
cets10
}
\description{
1000 time series following \code{cets9}.
}
\keyword{datasets}
