% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cets.R
\docType{data}
\name{cets12}
\alias{cets12}
\title{Data from CompEngine}
\format{An object of class \code{list} of length 1000.}
\usage{
cets12
}
\description{
1000 time series following \code{cets11}.
}
\keyword{datasets}
