% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compenginets-package.R
\docType{package}
\name{compenginets-package}
\alias{compenginets-package}
\alias{compenginets}
\title{Data from CompEngine}
\source{
\url{https://www.comp-engine.org/}.
}
\description{
The 25648 time series from CompEngine, a self-organizing database of time series data.
}
\author{
Rob J Hyndman. \email{Rob.Hyndman@monash.edu}

Yangzhuoran Yang. \email{Fin.Yang@monash.edu}
}
\keyword{package}
