% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cets.R
\name{get_cets}
\alias{get_cets}
\title{Extracting time series from the CompEngine database}
\usage{
get_cets(key)
}
\arguments{
\item{key}{A keyword describing attribute of the required time series}
}
\value{
A list consisting of the selected series.
}
\description{
Return time series where key matches some attribute
or is a logical vector indicating which rows of meta to return.
}
\examples{
cets_finance <- get_cets("finance")
}
\author{
Rob Hyndman
}
