% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\docType{data}
\name{meta}
\alias{meta}
\title{Metadata informarion of CompEngine database.}
\format{A data frame with 9 variables (the number of rows varies depending on the current size of CompEngine database):
\describe{
  \item{timeseries_id}{id of the time series}
  \item{timestamp_created}{time when the time series created}
  \item{source}{source of data measured}
  \item{category}{most specific possible category}
  \item{contributor}{creator who measured the data}
  \item{name}{name of the time series}
  \item{description}{detailed description of the data}
  \item{sampling_unit}{unit used in the time series}
  \item{sampling_rate}{measuring rate, assuming the data was measured from uniform sampling of some system through time}
}}
\source{
\url{https://www.comp-engine.org/#!browse}
}
\usage{
meta
}
\description{
A dataset containing the  attributes of \code{cets} data. Attributes include \code{category}, \code{timeseries_id},
\code{timestamp_created}, \code{source}, \code{contributor}, \code{name},
\code{description}, \code{sampling_unit}, and \code{sampling_rate}.
}
\keyword{datasets}
