% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compenginets-package.R
\docType{package}
\name{compenginets-package}
\alias{compenginets}
\alias{compenginets-package}
\title{compenginets: Time series from <https://www.comp-engine.org/>}
\description{
All time series from the comp-engine database are included, along with some meta data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkg.robjhyndman.com/compenginets/}
  \item \url{https://github.com/robjhyndman/compenginets}
  \item Report bugs at \url{https://github.com/robjhyndman/compenginets/issues}
}

}
\author{
\strong{Maintainer}: Rob Hyndman \email{Rob.Hyndman@monash.edu} (\href{https://orcid.org/0000-0002-2140-5352}{ORCID})

Authors:
\itemize{
  \item Yangzhuoran Yang \email{Fin.Yang@monash.edu}
}

}
\keyword{internal}
