% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cets.R
\name{get_cets}
\alias{get_cets}
\title{Extracting time series from the CompEngine database}
\usage{
get_cets(key, category = TRUE)
}
\arguments{
\item{key}{A keyword describing attribute of the required time series,
or a logical vector indicating which rows of meta to return.}

\item{category}{A logical value indicating whether the selection process
search \code{key} in \code{category} attribute. If \code{TRUE}, the \code{get_cets}
will return time series whose category match the keyword or belong to
some subcategory under the matched category. If \code{FALSE},
the function will return time series which have the keyword in their \code{timeseries_id},
\code{timestamp_created}, \code{source}, \code{contributor}, \code{name},
\code{description}, \code{sampling_unit}, or \code{sampling_rate} attributes.
This argument is only valid when \code{key} is not a logical vector.}
}
\value{
A list consisting of the selected series.
}
\description{
Return time series where key matches some attribute
or is a logical vector indicating which rows of meta to return.
}
\examples{
# Getting series within Finance category (including subcategory)
cets_finance <- get_cets("finance")
unique(mapply(attr, cets_finance, MoreArgs = list(which = "category")))

# Getting series whose sourse is Macaulay Library
cets_ML <- get_cets("Macaulay Library", category = FALSE)
unique(mapply(attr, cets_ML, MoreArgs = list(which = "source")))

# Extract time series by a random generated logical vector with length equals to the number of series
set.seed(2222)
idx <- sample(c(TRUE, FALSE), NROW(meta), replace = TRUE)
cets_logic <- get_cets(idx)
}
\author{
Rob J Hyndman

Yangzhuoran Yang
}
