% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cets.R
\name{get_cets}
\alias{get_cets}
\title{Extracting time series from the CompEngine database}
\usage{
get_cets(key, category = TRUE, maxpage = 10)
}
\description{
Extracting time series from the CompEngine database
}
\examples{
finance_m4 <- get_cets("M4_W138_Finance_1", category = FALSE)
a <- get_cets("real")
}
