% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cets.R
\name{get_cets}
\alias{get_cets}
\title{Extracting time series from the CompEngine database}
\usage{
get_cets(key, category_only = TRUE)
}
\arguments{
\item{key}{A keyword describing attribute of the required time series.}

\item{category_only}{A logical value indicating whether the selection process
search \code{key} in \code{category} attribute only. If set to \code{FALSE},
the function will also search the keyword in \code{source}, \code{name},
\code{timeseries_id},  \code{contributor}, and \code{description} attributes.}
}
\value{
A list consisting of the selected series.
}
\description{
Return time series where key matches some attribute
or is a logical vector indicating which rows of meta to return.
}
\examples{
cets_finance <- get_cets("finance")
}
\author{
Rob Hyndman
}
